Create base fastapi server
FastAPI Module used to create asynchronous web server.
Future updates will plan on adding an nginx for reverse
proxy to manage multiple instances of the fast api server.
This update also adds an enviornment variable file,
that uses the Pydantic library for type management.
The fastAPI application runs on a uvicorn ASGI sever.
The root application return the PORT number of that
the server is currently running in so we can track
replies from NGINX to visualize the load balancing.